requirements.txt
# Core data processing and analysis
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Web framework and API
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0

# Trading and market data
alpaca-py==0.40.1
TA-Lib==0.4.28
tiingo==0.1.0
financialmodelingprep==0.0.1

# Document processing
docx2pdf
python-docx
beautifulsoup4==4.12.2

# AI and analysis
openai==1.12.0 